Alterações nos codigos

+++ array.util.h
@@ -1,8 +1,8 @@
-int* allocateArray(int *array, long long int size);
+int* allocateArray(int *array, int size);
 void destructArray(int *array);
 
 //Printa na saída padrão os elementos de um array entre colchetes com os elementos separados por vírgula
-void printArray(int *array, long long int size);
+void printArray(int *array, int size);
 
 // Ordenação através do método shellSort
 void shellSort(int *a, int n);
